use flistxattr() and fgetxattr() when possible
When indexing xattrs, l[get|list]xattr() uses complete paths
to query xattrs for files. However, this creates a large overhead in
pathname resolution.

When possible, use f[list|get]xattr().

this is not always possible, for example, I do now know of a way to
make this work to query xattrs on a symbolic link.