debug mode: only dump queue size stats when exceeding highwater mark
In order to make the queue size stats more useful, and not be
overwhelemd with too much output, this change makes it so that queue
sizes are only printed when they exceed the previous highwater mark.

Also, this adds a slight code cleanup by removing the `#if defined`
around the callsite for `dump_queue_size_stats()`, and the function is
defined as a no-op when `QPTPOOL_QUEUE_SIZE` is undefined.